EditSearchResultFormを追加して適用
